Move input file parsing to config.go
