+++ tools/batchTestGenerator/internal/config.go
@@ -1,5 +1,11 @@
 package internal
 
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+)
+
 type RunConfig struct {
 	OutputLocation   string
 	IncludedServices map[string]struct{}
@@ -7,6 +13,36 @@ type RunConfig struct {
 	TestCaseFilePath string
 	EksARM64Vars     string
 	EksVars          string
+	TestCaseInput    Tests
+}
+
+type Tests struct {
+	Tests []Test `json:"tests"`
+}
+
+type Test struct {
+	CaseName  string   `json:"case_name"`
+	Platforms []string `json:"platforms"`
+}
+
+type TestCaseInfo struct {
+	testcaseName  string
+	serviceType   string
+	additionalVar string
+}
+
+func (r *RunConfig) ParseInputFile() error {
+
+	testCaseFile, err := os.ReadFile(r.TestCaseFilePath)
+	if err != nil {
+		return fmt.Errorf("failed to read test cases file: %w", err)
+	}
+
+	err = json.Unmarshal(testCaseFile, &r.TestCaseInput)
+	if err != nil {
+		return fmt.Errorf("failed to unmarshal test case file: %w", err)
+	}
+	return nil
 }
 
 func NewDefaultRunConfig() RunConfig {
